Add a waitForJsprMessage
While working in the firmware update it is useful to wait for a JSPR message
with specific return code. I.e. waiting for a 299 operationalState

This function should do this, it uses common delays that should be compatible
with Linux, Mac, Windows and Arduino.

Testing
-------
Just testing this implementation on Linux for now.