Improved the comments around SERIAL_CONTEXT_SETUP_FUNC macro
It was possibly unclear how this macro should be used and it was not surrounded
with #ifndef. To improve this a doxygen comment has been added and moved
to the header file to hopefully be more clear.

This macro is now surrounded with a #ifndef so users of the library can
define SERIAL_CONTEXT_SETUP_FUNC before including the library header.

Should there not be a SERIAL_CONTEXT_SETUP_FUNC definition the #ifndef in
the source file should raise a compile error.