Ensure rbBegin handles serial port reference safely (#24)
* Ensure rbBegin handles serial port reference safely

Previously, rbBegin accepted non const a pointer to the serial port and
did not take ownership of it. While this generally caused no immediate
issues—since the port was opened within rbBegin before the calling code
had a chance to modify or free the pointer—this was still not ideal
practice.

This change marks the serial port pointer as const to reflect the intended
usage and ensures the reference remains valid and unmodified during the
serial context usage.

This was just an oversight.

Testing
-------
Testing has only been done on Linux for the moment, further testing will
be done once pushed and updated in the PR.

* Fix windows build

There was a trailing semicolon in serial_windows breaking the build

Testing
-------
Did a full windows build and and got the hardware info a RB9704

* Fix arduino baud assignment

There was typo which was breaking the arduino build.

* Updated Doxygen documentation